% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{data}
\name{bdtnp}
\alias{bdtnp}
\title{Expression of 84 landmark genes in the Drosophila embryo}
\format{A numeric matrix with genes in columns and cells in rows, 84 columns
and 3039 rows.}
\source{
\url{http://bdtnp.lbl.gov:8080/Fly-Net/bioimaging.jsp?w=releases}
}
\usage{
bdtnp
}
\description{
The [Berkeley Drosophila Transcription Network Project (BDTNP)](http://bdtnp.lbl.gov:8080/Fly-Net/)
quantified expression of landmark genes at single cell resolution in
_Drosophila melanogaster_ embryos at different stages of development. This
dataset has BDTNP data for 84 landmark genes at late stage 5. The order of
cells in this dataset corresponds to the order of cells in the \code{\link{locations}}
dataset.
}
\keyword{datasets}
