% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_graph_dist.R
\name{.calc_graph_dist}
\alias{.calc_graph_dist}
\title{Internal function for graph-based distance}
\usage{
.calc_graph_dist(X, k, BNINDEX, BNPARAM, BPPARAM = SerialParam(), ...)
}
\arguments{
\item{X}{A numeric data matrix where rows are points and columns are dimensions.
        This can be missing if \code{BNINDEX} is supplied.}

\item{k}{An integer scalar for the number of nearest neighbors.}

\item{BNINDEX}{A \linkS4class{BiocNeighborIndex} object containing precomputed index information.
        This can be missing if \code{BNPARAM} is supplied, see Details.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the algorithm to use.
        This can be missing if \code{BNINDEX} is supplied, see Details.}

\item{BPPARAM}{a \linkS4class{BiocParallelParam} class for parallelization.}

\item{\dots}{Further arguments to pass to \code{\link[BiocNeighbors]{findKNN}}.}
}
\description{
This function is internal and is called by various S4 methods after method
specific pre-processing of data.
}
