% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{data}
\name{locations}
\alias{locations}
\title{Locations of cells in the Drosophila embryo}
\format{A data frame with 3 columns, xcoord, ycoord, and zcoord, and 3039 rows.}
\source{
\url{http://bdtnp.lbl.gov:8080/Fly-Net/bioimaging.jsp?w=releases}
}
\usage{
locations
}
\description{
This dataset has the x, y, and z spatial coordinates for cells in an
archetypical late stage 5 Drosophila embryo. The order of cells here is the
same as in the \code{\link{bdtnp}} dataset. This is only half of the embryo,
which is symmetric about the y axis here.
}
\keyword{datasets}
