% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spatial_expression}
\alias{plot_spatial_expression}
\title{Plot spatial gene expressions}
\usage{
plot_spatial_expression(gene_expressions, locations, symmetry,
  transposed = FALSE, interactive = FALSE, n_col = 2, pt_size = 1,
  alpha = 1)
}
\arguments{
\item{gene_expressions}{Numeric vector or matrix with genes in rows and
locations in columns. Matrices with genes in columns instead are also permitted,
with \code{transposed = TRUE}.}

\item{locations}{Data frame or numeric matrix with 1 to 3 columns specifying
locations, assumed to have x in the first column, y in the second, and z in
the 3rd.}

\item{symmetry}{A string for the column name in \code{locations} that specifies
an axis about which symmetry is present. This is useful if due to symmetry,
only half of the locations are in \code{locations} as the other half would be
the same anyway. However, if all locations are present in \code{locations}
even though the actual tissue does have symmetry, this argument can be left
missing.}

\item{transposed}{Logical, whether the matrix in \code{gene_expressions} has
genes in columns and locations in rows. Defaults to \code{FALSE}.}

\item{interactive}{Logical, whether the plot returned should be interactive.
If \code{TRUE}, then a Plotly-based html widget will be returned. Otherwise
a \code{ggplot} object. Defaults to \code{FALSE}. When there are 3 columns
in \code{locations}, this argument will be ignored and an interactive plot
will always be returned.}

\item{n_col}{Number of columns in multi-faceted plot when multiple genes are
plotted at once.}

\item{pt_size}{Numeric, point size.}

\item{alpha}{Numeric, transparency of points.}
}
\value{
If \code{interactive = FALSE} and the plot is less than 3D, a \code{ggplot2}
object. Otherwise a Plotly object.
}
\description{
This function can be used to plot expression of one or more genes in spatial
locations. This function can be used to visualize predictions from this
package, and gene expression in general as well if locations are known.
Plotly-based interactive graphics can be generated, but by default, ggplot2
will be used.
}
