% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_expr_loc}
\alias{predict_expr_loc}
\title{Predict gene expression at locations}
\usage{
predict_expr_loc(X, cell_loc, transposed = FALSE, scale = TRUE)
}
\arguments{
\item{X}{Numeric matrix (can be sparse) with genes in rows and cells in
columns. If genes are in columns, then set \code{transposed = TRUE}.}

\item{cell_loc}{Matrix that probabistically assigns cells to locations, made
from the function \code{\link{gw_assign}}.}

\item{transposed}{Logical, whether the matrix has cells in rows rather than
in columns. Defaults to \code{FALSE}.}

\item{scale}{Logical, whether predictions should be scaled so their means are
the same as in single cell RNA-seq data supplied in \code{x}.
Defaults to \code{TRUE}.}
}
\value{
A numeric matrix with genes in rows and locations in columns.
}
\description{
With the probabilistic assignment of cells to locations from optimal transport,
we can predict gene expression at each location based on the gene count matrix.
Let \eqn{X} denote the gene count matrix with genes in rows and cells in
columns, and \eqn{T} denote the probabilistic assignment of cells to locations,
with cells in rows and locations in columns. Gene expression at each location
can be predicted by \eqn{XT}. This function will use this to predict gene
expression. It will also scale the prediction so the mean gene expression
across locations will match the mean gene expression among cells in scRNA-seq.
}
